Chore: Rename `ws` -> `webrtc`
