+++ aether_server/routes/views/__init__.py
@@ -1,5 +1,5 @@
 from .index_view import AetherIndexView
-from .ws_view import AetherWebRTCView
+from .webrtc_view import AetherWebRTCView
 
 __all__ = [
     "AetherIndexView",
